feat: Add and structure new content in the blog section, Add visual adjustments and styling changes to blog and category sections
Blog page: Adjusted the layout of the blog hero section to align the text on the left and the image on the right. Each blog post item is now displayed in a stacked manner with images positioned to the right of the text.

All posts section: Added a horizontal line beneath the "All posts" heading to create separation and visually improve the layout. The heading was aligned to the left for consistency.

Category adjustments: Increased the size of the business.png image compared to other category icons to make the Business category more prominent. Appropriate SCSS and HTML code was added for this adjustment.

Fixes: General improvements to layout alignment and style consistency throughout the page. The structure has been optimized following the BEM methodology.